refactor(oas): move XML serializer from openapi-sampler
closes #119
