+++ packages/oas/package.json
@@ -47,6 +47,7 @@
     "@har-sdk/core": "*",
     "@har-sdk/openapi-sampler": "*",
     "json-pointer": "^0.6.2",
+    "jstoxml": "^3.2.6",
     "qs": "^6.10.3",
     "tslib": "^2.3.1",
     "url-template": "^2.0.8"
@@ -54,6 +55,7 @@
   "devDependencies": {
     "@types/js-yaml": "^4.0.5",
     "@types/json-pointer": "^1.0.31",
+    "@types/jstoxml": "^2.0.2",
     "@types/qs": "^6.9.7",
     "@types/url-template": "^2.0.28",
     "js-yaml": "^4.1.0"
