+++ packages/oas/src/converter/index.ts
@@ -1,5 +1,6 @@
 export * from './parts';
 export * from './security';
+export * from './serializers';
 export * from './Converter';
 export * from './DefaultConverter';
 export * from './Sampler';
