+++ packages/oas/src/converter/parts/postdata/BodyConverter.ts
@@ -64,6 +64,10 @@ export abstract class BodyConverter<T extends OpenAPI.Document>
         return this.encodeJson(value);
       case 'application/x-www-form-urlencoded':
         return this.encodeFormUrlencoded(value);
+      case 'application/xml':
+      case 'text/xml':
+      case 'application/atom+xml':
+        return this.encodeXml(value);
       case 'multipart/form-data':
       case 'multipart/mixed':
         return this.encodeMultipartFormData(value, encoding);
@@ -166,6 +170,11 @@ export abstract class BodyConverter<T extends OpenAPI.Document>
       format: 'RFC3986',
       encode: false
     });
+  }
+
+  // TODO: implement the method using {@link XmlSerializer}
+  private encodeXml(_: unknown): string {
+    return `<?xml version="1.0" encoding="UTF-8"?>`;
   }
 
   private encodeOther(value: unknown): string {
