+++ packages/oas/src/converter/serializers/Serializer.ts
@@ -1,0 +1,5 @@
+import { OpenAPIV3 } from '@har-sdk/core';
+
+export interface Serializer {
+  serialize(data: unknown, schema: OpenAPIV3.SchemaObject): string;
+}
