+++ packages/openapi-sampler/package.json
@@ -44,12 +44,10 @@
   "dependencies": {
     "@har-sdk/core": "*",
     "json-pointer": "^0.6.2",
-    "jstoxml": "^3.2.6",
     "randexp": "^0.5.3",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@types/json-pointer": "^1.0.31",
-    "@types/jstoxml": "^2.0.2"
+    "@types/json-pointer": "^1.0.31"
   }
 }
