+++ packages/openapi-sampler/src/index.ts
@@ -4,12 +4,10 @@ import {
   NullSampler,
   NumberSampler,
   ObjectSampler,
-  OpenAPISchema,
   Sampler,
   StringSampler
 } from './samplers';
 import { DefaultTraverse, Options, Schema, Specification } from './traverse';
-import { XmlSerializer } from './serializers';
 
 export const SAMPLER_MAP: Map<string, Sampler> = new Map();
 
@@ -34,16 +32,10 @@ export const sample = (
   SAMPLER_MAP.set('object', new ObjectSampler(traverse));
   SAMPLER_MAP.set('string', new StringSampler());
 
-  const xml = new XmlSerializer();
-
   traverse.samplers = SAMPLER_MAP;
   traverse.clearCache();
 
-  const { value } = traverse.traverse(schema, opts, spec) ?? {};
-
-  return typeof value === 'object' && options?.serializeToXml
-    ? xml.serialize(value, schema as OpenAPISchema)
-    : value;
+  return traverse.traverse(schema, opts, spec)?.value;
 };
 
 export { Schema, Options, Specification } from './traverse';
