--- packages/openapi-sampler/src/serializers/Serializer.ts
@@ -1,5 +1,0 @@
-import { OpenAPISchema } from '../samplers';
-
-export interface Serializer {
-  serialize(data: unknown, schema: OpenAPISchema): string;
-}
